Consolidated abstract test classes
